🚚 [RUM-253] move some deflate types to core
