+++ packages/rum/src/domain/deflate/deflateWorker.ts
@@ -1,4 +1,4 @@
-import type { DeflateWorkerAction, DeflateWorkerResponse } from '@datadog/browser-core'
+import type { DeflateWorker, DeflateWorkerResponse } from '@datadog/browser-core'
 import { addTelemetryError, display, includes, addEventListener, setTimeout, ONE_SECOND } from '@datadog/browser-core'
 import type { RumConfiguration } from '@datadog/browser-rum-core'
 
@@ -37,10 +37,6 @@ type DeflateWorkerState =
       stop: () => void
       version: string
     }
-
-export interface DeflateWorker extends Worker {
-  postMessage(message: DeflateWorkerAction): void
-}
 
 export type CreateDeflateWorker = typeof createDeflateWorker
