+++ packages/rum/src/domain/deflate/index.ts
@@ -1,6 +1,5 @@
-export { DeflateEncoderStreamId, DeflateEncoder, createDeflateEncoder } from './deflateEncoder'
+export { createDeflateEncoder } from './deflateEncoder'
 export {
-  DeflateWorker,
   startDeflateWorker,
   DeflateWorkerStatus,
   getDeflateWorkerStatus,
