+++ packages/rum/src/domain/segmentCollection/segment.spec.ts
@@ -1,12 +1,11 @@
-import type { TimeStamp } from '@datadog/browser-core'
-import { noop, setDebugMode, isIE } from '@datadog/browser-core'
+import type { DeflateEncoder, TimeStamp } from '@datadog/browser-core'
+import { noop, setDebugMode, isIE, DeflateEncoderStreamId } from '@datadog/browser-core'
 import type { RumConfiguration } from '@datadog/browser-rum-core'
 import { MockWorker } from '../../../test'
 import type { CreationReason, BrowserRecord, SegmentContext, BrowserSegment, BrowserSegmentMetadata } from '../../types'
 import { RecordType } from '../../types'
 import { getReplayStats, resetReplayStats } from '../replayStats'
-import type { DeflateEncoder } from '../deflate'
-import { DeflateEncoderStreamId, createDeflateEncoder } from '../deflate'
+import { createDeflateEncoder } from '../deflate'
 import type { AddRecordCallback, FlushCallback } from './segment'
 import { Segment } from './segment'
